Cleaned up code (utils.js & Fade.js)
